update deleted files
