+++ .gitignore
@@ -125,3 +125,4 @@ settings.json
 
 # other
 *.pyc
+*.xml
