--- r.txt
@@ -1,10 +1,0 @@
-[
-{"author": 3, "channel": 4, "message": "Hi", "datetime": 3},
-{"author": 3, "channel": 4, "message": "Hi", "datetime": 3},
-{"author": 3, "channel": 4, "message": "Hi", "datetime": 3},
-{"author": 3, "channel": 4, "message": "Hi", "datetime": 3},
-{"author": 3, "channel": 4, "message": "Hi", "datetime": 3},
-{"author": 3, "channel": 4, "message": "Hi", "datetime": 3},
-{"author": 3, "channel": 4, "message": "Hi", "datetime": 3}
-]
-"a"
